SI-7779 Account for class name compactification in reflection
We have to assume that the classes we are reflecting on were
compiled with the default value for -Xmax-classfile-name (255).

With this assumption, we can apply the same name compactification
as done in the regular compiler.

The REPL is particularly prone to generating long class names
with the '$iw' prefixes, so this is an important fix for runtime
reflection.

Also adds support for getting the runtime class of `O.type` if
`O` is a module.

(cherry picked from commit 5dbc37dfbe0e9a039da6744e45012abc3034cdf5)
[backport from 2.10.x to 2.10.3]