Annotate test sources with pattern inference data.
Attempting to innovate in the field of not squandering all the
hard-won knowledge from days of fiddling with the implementation,
I devised this means of making the next guy's life easier. In
this commit certain tests are annotated with comments elaborating
on how a pattern's type was inferred. The annotated tests meet
these conditions:

 - they contain a "Typed" pattern (e.g. x: Foo => ... )
 - the candidates for inference are not type-equivalent

The type of a "Typed" pattern is inferred based on the
expected type of the pattern (the type of the "scrutinee") and
the ascripted type. These types are known internally as "pt" and
"pattp", respectively. The essential candidates for the inferred
type are:

  intersect(pattp, pt)
  intersect(pt, pattp)

From this apparently simple choice arises great complication,
arising in part from the language semantics - intersections are
asymmetric - and in part from the implementation - the mutation
based approach to type inference makes it extremely difficult
to modify, e.g. simply by performing a subtype test you stand a
good chance of changing the outcome. Furthermore one faces a mix
of properly formed types and types carrying dummy type arguments,
TypeVars, Wildcards, and other constructs which don't always lead
to intuitive results for <:<.

The tiniest, most apparently irrelevant change to the key method
inferTypedPattern is(was) liable to cause a test to change behavior.
For this reason the most valuable part of this commit may be simply
a side effect of annotating all these tests at once: it is now very
easy to test these and only these, e.g.

  test/partest $(git show --pretty=format: --name-only $THIS_COMMIT_SHA)

A note on notation: the meanings of the conformance operators
are mostly self-evidence, but I had to invent two:

  a !:= b  // (a =:= b), (a <:< b), and (b <:< a) are all FALSE
  a ~:= b  // (a =:= b) is FALSE, but (a <:< b) and (b <:< a) are TRUE

That last is supposed to call to mind "approximately =:=". This
outcome, where a <:< b, b <:< a, and !(a =:= b) is actually common
at the positionm where it is being recorded.