SI-7472 superior pattern type inference.
Major cleanup in type inference. Lose less information about
the type of the match scrutinee: don't widen it arbitrarily
(so a singleton type can survive match, see pos/t6771b.scala)
and preserve it through the match when the straitjacket of an
implementation allows for it.

More generally, shone some light into the darkest corners of type
inference. It's still a grim apocalyptic wasteland, but maybe the
next guy will have a better chance.

The real contribution here is to bring a sliver of abstraction
to the creation and instantiation of TypeVars. They are still
a long way from encapsulated, but feel closer now. Should
someone wish to continue this effort (assuming continuing with
the current mechanism based on mutation of TypeVars, which I
would not recommend) I advise that nothing be allowed to create
TypeVars directly. Every use of TypeVars should be encapsulated
such that the TypeVars are created on method entry and either
instantiated or discarded by method exit.