Make the DelegatingReporter aware of -nowarn
The test case compiles a project without and with this
setting and checks that a warning is and isn't emitted
respectively.

It's a multi-project build; this bug didn't seem to turn
up in a single-project build.

Rewritten from sbt/zinc@75a6e92cc74d7b6e0644bcd2fad76c9aecf5167e