Avoid spurious tree checker warning for higher order type params
TreeCheckers is trying to find references to a) local types
(term owned) or b) to type parameters from a trees that are
not ancestors of the a) term or b) type param owner. Such
references are ill-scoped and suggest that a tree has been
transplanted without proper substitution.

However, this check failed to account for higher order type
parameters, as seen in the spurious warning emitted by:

    test/pending/pos/treecheckers/c5.scala.