Build partest-extras under `pack.done`
... rather than just in `test.suite.init`.

Now:

	% ant pack.done | egrep -i 'compiling|jar'
	desired.jars.uptodate:
	[quick.partest-extras] Compiling 1 file to /Users/jason/code/scala/build/quick/classes/partest-extras
	      [jar] Building jar: /Users/jason/code/scala/build/pack/lib/scala-partest-extras.jar

Note: Because of the recent changes to the way that the classpath
or partest is build up (it is done via `ant test.suite.init`), partest
no longer works with quick/classes, the classpath is always taken as
`pack`. So `ant quick.bin && ./test/partest` is insufficient; you need
to run `ant pack.done`, or just `ant` if you prefer brevity.