Fix encoding problems in output of scaladoc -help
From the output of `scaladoc -help`:

```
  -doc-source-url <url>                       A URL pattern used to
  build links to template sources; use variables, for example:
  ?{TPL_NAME} ('Seq'), ?{TPL_OWNER} ('scala.collection'),
  ?{FILE_PATH} ('scala/collection/Seq')
```

The ? should be €. Fix that in the source, using the correct
encoding (which is already UTF-8). This requires the *default JVM
encoding* to be UTF-8 when the script is run, that is, passing
`-Dfile.encoding=UTF-8`.

So, after this commit, the euro sign is garbled in the output of
`./qbin/scaladoc -help` and correctly displayed in the output of
`./qbin/scaladoc -Dfile.encoding=UTF-8 -help`.

The next step is to make that option the default, but of course that
change has more potential for breakage.

The encoding problem was introduced in 2011 by commit
2621ee63285808785159a3c24c9e5a5a723b8b9c, and seems to not have been
reported till now.