Scala tools: Set default IO encoding to UTF-8
This changes the template for tool runners to pass
`-Dfile.encoding=UTF-8`. The necessity was uncovered by previous commit.

Alternatively, `-Dfile.encoding=UTF-8` might need to be part of the
local setup - in which case the previous commit is sufficient to fix
Scaladoc's bug.

This is mostly a request for comments; I suppose we want this
change (eventually), and this is the simplest way to change the default
encoding, but I'm not sure it's robust enough.

Moreover, it will break on non-UTF-8 consoles. But since the only
non-ASCII console output seems to be three characters in `scaladoc
-help`'s output, and they've been broken for everybody for two years,
this *might* be acceptable.

The setting is currently not passed again as argument to Scala, and is
not overridable by the user.

My testing was:

* run:
```sh
rm build/quick/bin.complete
ant quick.bin
./qbin/scaladoc -help
```

* observe the output is now not garbled.

I suppose this does not affect the actual input or output of Scala
tools, since they have a separate `-encoding` option to specify that.

The testing I've done is clearly insufficient, and I fear the testsuite
will also not test this aspect (since launchers are usually bypassed).
Testing on Windows is also necessary.