Initial work on flat classpath
The goal of these changes is to create an alternative classpath
implementation which would be possibly more efficient and would
use less memory. It can be enabled using -YclasspathImpl:flat flag.
This implementation is more low level than the old one as e.g. .NET
backend has been dropped so we can perform operations directly on
a few Java-specific file types.

It was needed to abstract over the new and the old implementation.

There's added hack for sbt incremental compiler interface. It depends
on ClassPath.findClass method. Since there's no common interface between
the old ClassPath and the new FlatClasspath, the old one has been changed
to be aware of the new one, and dynamically dispatch to FlatClasspath if
it's enabled. The old classpath is all lazy so if the only call to the old
classpath is coming from incremental compiler it'll be dispatched to the
new classpath and nothing in old classpath will get forced.