SI-7843 Restore JSR 223 service entry
The 2.10 fix to remove the ScriptEngine service entry
was inadvertently forwarded to 2.11.

This commit reverts and adds a test.

This situation was entirely foreseen by retronym,
proving beyond doubt that he is in fact a time traveler,
as hinted by his name. He brings bugs forward into the
future and returns into the past with fixes and other
alien technology like scalaz.