Move logic checking valid names from ClassPath to ClassPathContext
Other methods taking the same kind of decisions were already
in ClassPathContext, e.g., isValidName() or, in some sense,
even toBinaryName().

This makes ClassPath itself be completely agnostic of how
particular kinds of files or directories are named.

It also allows to override this logic at the context level.
Without it, overriding this logic required a fair amount of
code duplication from ClassPath.