Don't avoid to load trait impl .class without inliner.
When the inliner is on, it is necessary to load class
files containing the implementation of traits, i.e.,
those ending in $class.class. This is needed to be
able to inline these methods. However it is useless
if the inliner is disabled.

The previous logic avoided to put these files on the
classpath when the inliner was off. However, this
complicates things, and it makes the classpath
mechanism dependent on something totally unrelated.

On this basis, this commit removes that logic, and
instead trait impl .class files are always kept on
the classpath.