SI-7870 Detect default getter clashes in constructors
Default getters for constructors live in the companion module.
These eluded the check for clashes in default getter names due
to overloading, which aims to give a more user friendly error
than "double definition: meth$default$1".

This commit checks for default getters in the companion module,
in addition to those in the template itself.