SI-4742 Make -Xcheckinit aware of constants.
Members defined as `final val x = <literal>` are given
a ConstantType. The constant is folded into the accessor
method `x`, and the field itself is never initialized.
(Related discussion: SI-4605)

As such, -Xcheckinit spuriously warns when calling that
accessor.

This commit disables the checks for constants.

This will also fix the checkinit build (failure tracked as SI-7839),
which is the victim of this a spurious scolding.