SI-7876 Less dealiasing in Scaladoc's type printing.
Scaladoc renders tuple and function types with the syntactic sugar.

To do so, it was using `isFunctionType` and `isTupleType`. Internally,
these methods dealias, so one can't simply take the type arguments
from the given type to get the element types of the tuple.

In ac96200c92, a Scaladoc crasher in this area was resolved by
normalizing a type before extracting Tuple type arguments.

Similar code already existed in the handling of FunctionTypes.

This commit goes in the opposite direction, and instead queries
the `direct` versions of those methods, which do not perform
dealiasing.

Conseqeuently, the type aliases that appear in the source code
signatures for members will be rendered as such in Scaladoc.
Which seems like an improvement.