refactors ToolBoxGlobal wrappers
Centralizes wrappers in a single location and makes it impossible to use
the compiler without going through that location.

Also fixes the option string tokenization bug that was there for ages.
This time I figured out that there's already an implementation of the
tokenizer in our compiler, so I just used it :)